Address Pub/Sub review compatibility gaps
